requirements.txt
selenium
appium-python-client
pytest
pytest-html
python-dotenv
faker